FEM: move modules MechanicalMaterial to a more general FemMaterial
