+++ src/Mod/Fem/Gui/CMakeLists.txt
@@ -344,7 +344,6 @@ fc_target_copy_resource(FemGui
     ${CMAKE_SOURCE_DIR}/src/Mod/Fem
     ${CMAKE_BINARY_DIR}/Mod/Fem
     InitGui.py
-    TaskPanelMechanicalMaterial.ui
     )
 
 SET(FemGuiIcon_SVG
