+++ src/Mod/Fem/Gui/Workbench.cpp
@@ -73,7 +73,7 @@ Gui::ToolBarItem* Workbench::setupToolBars() const
          << "Fem_MeshRegion"
          //<< "Fem_CreateNodesSet"
          << "Separator"
-         << "Fem_MechanicalMaterial"
+         << "Fem_Material"
          << "Fem_MaterialMechanicalNonlinear"
          << "Fem_BeamSection"
          << "Fem_ShellThickness"
@@ -145,7 +145,7 @@ Gui::MenuItem* Workbench::setupMenuBar() const
          << "Fem_MeshRegion"
          << "Fem_CreateNodesSet"
          << "Separator"
-         << "Fem_MechanicalMaterial"
+         << "Fem_Material"
          << "Fem_MaterialMechanicalNonlinear"
          << "Fem_BeamSection"
          << "Fem_ShellThickness"
